// frontend/src/app/api/model/sloidLocationModel.ts
/**
 * Atlas API
 *
 * Contact: TechSupport-ATLAS@sbb.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SloidLocationModel { 
    sloid?: string;
    sloidType?: SloidLocationModel.SloidTypeEnum;
}
export namespace SloidLocationModel {
    export type SloidTypeEnum = 'SERVICE_POINT' | 'AREA' | 'PLATFORM' | 'REFERENCE_POINT' | 'TOILET' | 'PARKING_LOT' | 'CONTACT_POINT' | 'SECTOR' | 'SECTOR_GROUP';
    export const SloidTypeEnum = {
        ServicePoint: 'SERVICE_POINT' as SloidTypeEnum,
        Area: 'AREA' as SloidTypeEnum,
        Platform: 'PLATFORM' as SloidTypeEnum,
        ReferencePoint: 'REFERENCE_POINT' as SloidTypeEnum,
        Toilet: 'TOILET' as SloidTypeEnum,
        ParkingLot: 'PARKING_LOT' as SloidTypeEnum,
        ContactPoint: 'CONTACT_POINT' as SloidTypeEnum,
        Sector: 'SECTOR' as SloidTypeEnum,
        SectorGroup: 'SECTOR_GROUP' as SloidTypeEnum
    };
}